Add CloudFormation stack for monitoring
We want to have a controlled way of rolling out updates to monitoring
and metrics collection.

This adds:

* alarms based on metrics hitting particular thresholds
* a dashboard for the Advice service

We are defining this as Infrastructure as Code so that we can try out
new things in a lower environment. UAT is our production-like
environment.

Once we are confident that a change is good in our lower environment, we
can apply the same change to production.

By defining it in code, we can use normal iteration and review
processes to get to a known good state. We can then deploy the same
changes in production and be confident that we have not mistyped/
misconfigured a change in production.